tada68: Ensure PCB does not hit dove tails
Define the area below the PCB which should not be touched by any objects
from below. With that, ensure that the locking mechanism does not
interfere with the PCB.
Additionally, move the stand_height variable to the global space and
increment its value from 4.5 to 5 to allow a minimal rim around the
locking mechanism.

Fix the missing dependency of the USB cutout translation (it should
depend on stand_h).